requirements.txt
groq
phidata
streamlit
duckduckgo-search
nest_asyncio
newspaper4k
lxml_html_clean